+++ src/transport/xgl_reliable.c
@@ -98,8 +98,8 @@ void xgl_reliable_destroy(xgl_reliable_queue_t* queue) {
 xgl_error_t xgl_reliable_add_packet(xgl_reliable_queue_t* queue,
                                     const uint8_t* data,
                                     size_t data_len,
-                                    uint8_t source_id,
-                                    uint8_t target_id,
+                                    uint16_t source_id,
+                                    uint16_t target_id,
                                     uint8_t seq_num,
                                     uint8_t data_type,
                                     uint8_t priority,
@@ -162,7 +162,7 @@ xgl_error_t xgl_reliable_add_packet(xgl_reliable_queue_t* queue,
  */
 xgl_error_t xgl_reliable_remove_packet(xgl_reliable_queue_t* queue,
                                        uint8_t seq_num,
-                                       uint8_t target_id) {
+                                       uint16_t target_id) {
     if (queue == NULL) {
         return XGL_ERR_NULL_POINTER;
     }
@@ -301,7 +301,7 @@ void xgl_reliable_clear(xgl_reliable_queue_t* queue) {
  */
 xgl_reliable_packet_t* xgl_reliable_find_packet(const xgl_reliable_queue_t* queue,
                                                 uint8_t seq_num,
-                                                uint8_t target_id) {
+                                                uint16_t target_id) {
     if (queue == NULL) {
         return NULL;
     }
